requirements.txt
python-chess
numpy
art
Pillow
requests
vapi-python
opencv-python